Main/BackEnd/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn
psycopg2-binary
dj-database-url
whitenoise